Add `GET /api/v3/reminders` endpoint (#19083)
An endpoint that lists all reminders for a user, regardless of the workpackage.
