+++ lib/api/v3/work_packages/work_packages_api.rb
@@ -98,7 +98,7 @@ module API
             mount ::API::V3::WorkPackages::AvailableProjectsOnEditAPI
             mount ::API::V3::WorkPackages::AvailableRelationCandidatesAPI
             mount ::API::V3::WorkPackages::WorkPackageRelationsAPI
-            mount ::API::V3::Reminders::RemindersAPI
+            mount ::API::V3::Reminders::RemindersByWorkPackageAPI
             mount ::API::V3::EmojiReactions::EmojiReactionsByWorkPackageCommentsAPI
           end
